refactor: Activity, Project, SectionTitle 스타일 분리
